# Use RARRAY_AREF() instead to improve array accessing