--- a/ext/oj/wab.c
+++ b/ext/oj/wab.c
@@ -128,7 +128,7 @@ static void dump_array(VALUE a, int depth, Out out, bool as_ok) {
         for (i = 0; i <= cnt; i++) {
             assure_size(out, size);
             fill_indent(out, d2);
-            oj_dump_wab_val(rb_ary_entry(a, i), d2, out);
+            oj_dump_wab_val(RARRAY_AREF(a, i), d2, out);
             if (i < cnt) {
                 *out->cur++ = ',';
             }